+++ test/unit/fuzz-tools.test.js
@@ -109,6 +109,7 @@ import * as eddecisionV107 from '../../lib/eddecision-v107.js';
 import * as traumaV108 from '../../lib/trauma-v108.js';
 import * as traumaclassV109 from '../../lib/traumaclass-v109.js';
 import * as toxV110 from '../../lib/tox-v110.js';
+import * as enviroV111 from '../../lib/enviro-v111.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -162,6 +163,7 @@ const MODULES = {
   'trauma-v108.js': traumaV108,
   'traumaclass-v109.js': traumaclassV109,
   'tox-v110.js': toxV110,
+  'enviro-v111.js': enviroV111,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];